Specialize plot and plot3d functions
