+++ src/poliastro/plotting/__init__.py
@@ -1,2 +1,2 @@
-from .interactive import OrbitPlotter2D, OrbitPlotter3D
-from .common import plot, plot3d, plot_solar_system
+from .interactive import OrbitPlotter2D, OrbitPlotter3D, plot, plot3d
+from .common import plot_solar_system
